Trim mdlManager 00314850 third local width

+++ src/Graphics/Model/mdlFile.c
@@ -425,7 +425,7 @@ int FUN_00326240(int param_1,int param_2);
 void FUN_00326280(int param_1,u64 param_2,u8 (*param_3) [16]);
 void FUN_003263b0(int param_1,u64 param_2,u8 (*param_3) [16]);
 u32 FUN_00326490(u32 param_1);
-u64 FUN_00326510(u64 param_1);
+u64 FUN_00326510(u32 param_1);
 void FUN_00326740(int param_1);
 u64 FUN_003268c0(u64 param_1);
 void FUN_003269e0(int param_1,int param_2);
@@ -10271,7 +10271,7 @@ u32 FUN_00326490(u32 param_1)
 // FUN_00326510 NONMATCHING
 
 
-u64 FUN_00326510(u64 param_1)
+u64 FUN_00326510(u32 param_1)
