+++ src/Graphics/Model/mdlManager.c
@@ -4176,7 +4176,7 @@ void func_00314850(RpClump* param_1,void* param_2,s16 param_3,u16 param_4,u16 pa
 
   u32 lVar7;
 
-  long lVar8;
+  u32 lVar8;
 
   u16 *puVar9;
